Modified: BugFix: Checking the branches of new reconstructed tree.
BUgDes: Some reconstruction method like Bionj in test 3 file make the branch length of size zero. This is not possible since later for likelihood computation it is problematic.

BugFix: Now, we check the branch and in the case of more than one branches with zero length, we will ask user to change the tree reconstruction method.