Modified: BugFix: Resolve inferring ambiguous 'M'/'B' character as ancestral state
BugDes: During the inference in when internal node were chosen as the new root of the pruned tree and only has one child and that one child has only one and so on: the implemented BPP functions 'getRootId' and 'isLeaf' have returned a wrong value for the new root. Meaning since the new root has only one child (for the leaf detection it counts the number of neighbours), the isLeaf returns true and getRootId did not work as it suppose to. This only happened for the new root which as internal node has the corresponding function to call the reconstruction phase of algorithm. As a result, the values of new root and internal nodes between the new root and leaves did not have value. Please note that this issue were addressed before using the condition to check the root id.

BugFix: by double checking the inconsistency of node's id and newRootId, we make sure that the new root would not be treated as leaf.