BugFix: Set the branch length to a very small value (1e-8) for Branch length zero.
BugDes: This is one of the cases that user forget to make sure that the branch length of zero is not valid except for the root. So we have to check it here, resolve it and warn the user.

BugFix: To resolve it, we set the branch length to a very small value (1e-8).